+++ Classification/include/CGAL/Classification/Feature/Echo_scatter.h
@@ -55,7 +55,7 @@ public:
 private:  
   typedef Classification::Image<float> Image_float;
 
-  std::vector<double> echo_scatter;
+  std::vector<float> echo_scatter;
   
 public:
   /*!
@@ -70,8 +70,8 @@ public:
   Echo_scatter (const PointRange& input,
                 EchoMap echo_map,
                 const Grid& grid,
-                const double grid_resolution,
-                double radius_neighbors = 1.)
+                const float grid_resolution,
+                float radius_neighbors = 1.)
   {
     this->set_name ("echo_scatter");
     Image_float Scatter(grid.width(), grid.height());
@@ -97,7 +97,7 @@ public:
           for(std::size_t k = squareXmin; k <= squareXmax; k++){
             for(std::size_t l = squareYmin; l <= squareYmax; l++){
 									
-              if(CGAL::sqrt(pow((double)k-i,2)+pow((double)l-j,2))<=(double)0.5*radius_neighbors/grid_resolution){
+              if(CGAL::sqrt(pow((float)k-i,2)+pow((float)l-j,2))<=(float)0.5*radius_neighbors/grid_resolution){
 										
                 if(grid.indices(k,l).size()>0){
 									
@@ -128,12 +128,12 @@ public:
     for(std::size_t i = 0; i < input.size(); i++){
       std::size_t I= grid.x(i);
       std::size_t J= grid.y(i);
-      echo_scatter.push_back((double)Scatter(I,J));
+      echo_scatter.push_back((float)Scatter(I,J));
     }
   }
 
   /// \cond SKIP_IN_MANUAL
-  virtual double value (std::size_t pt_index)
+  virtual float value (std::size_t pt_index)
   {
     return echo_scatter[pt_index];
   }
